requirements.txt
# RAG-LLM Dependencies
# Обновлено: 2026-03-24

# === Backend ===
fastapi>=0.135.0            # SSE support, Pydantic V3 integration
uvicorn[standard]>=0.34.0   # ASGI сервер с HTTP/2
pydantic>=2.10.0            # 50x быстрее с Rust-backed core
pydantic-settings>=2.13.0   # Конфигурация через env vars
python-dotenv>=1.0.0        # Загрузка .env файлов

# === HTTP ===
httpx>=0.28.0               # Async HTTP клиент с HTTP/2

# === gRPC ===
grpcio>=1.62.0              # gRPC runtime
grpcio-tools>=1.62.0        # protoc компилятор для Python
grpcio-reflection>=1.62.0   # Server reflection (grpcurl discovery)
protobuf>=4.25.0            # Protocol Buffers

# === Парсинг документов ===
google-genai>=1.0.0           # Gemini SDK для парсинга PDF в облаке
docling>=2.80.0             # IBM: PDF/DOCX/PPTX парсинг (локально, optional)

# === Векторная БД ===
chromadb>=1.0.0             # v2 API, Rust backend, sparse vectors

# === Keyword search (BM25) ===
bm25s>=0.3.0                # Быстрый BM25 (scipy sparse)
snowballstemmer>=3.0.0      # Pure-Python Snowball стеммер (ru, en, ...)

# === Мониторинг ===
psutil>=6.1.0               # CPU, RAM, диск, сеть
pynvml>=12.0.0              # NVIDIA GPU метрики

# === Логирование ===
loguru>=0.7.3               # Структурное логирование

# === Утилиты ===
python-multipart>=0.0.18    # Загрузка файлов через multipart/form-data
PyMuPDF>=1.25.0             # PDF → PNG рендер для Vision LLM парсинга
PyYAML>=6.0                 # YAML-профили для embedding-провайдеров
